first pass at upsert for scheduled functions
