+++ src/create/aws/create-scheduled-lambda-deployments/readme.md
@@ -1,0 +1,7 @@
+## phase 1
+- [x] always putRule on `npx create` if a scheduled function (fixes #326)
+- [ ] enhanced cron() and rate() validatation (closes #148)
+
+phase 2
+- [ ] state enabled/disabled flag in .arc-config w `npx config` (if a scheduled function)
+- [ ] concurrency 0-1 flag  in `.arc-config` w `npx config` (all functions)
